Remove systemconfig
