+++ bundle/manifests/multiarch-tuning-operator.clusterserviceversion.yaml
@@ -27,7 +27,7 @@ metadata:
     categories: OpenShift Optional, Other
     console.openshift.io/disable-operand-delete: "false"
     containerImage: registry.ci.openshift.org/origin/multiarch-tuning-operator:main
-    createdAt: "2024-10-16T20:07:51Z"
+    createdAt: "2025-01-14T00:47:27Z"
     features.operators.openshift.io/cnf: "false"
     features.operators.openshift.io/cni: "false"
     features.operators.openshift.io/csi: "false"
@@ -139,16 +139,6 @@ spec:
           verbs:
           - get
         - apiGroups:
-          - config.openshift.io
-          resources:
-          - imagedigestmirrorsets
-          - images
-          - imagetagmirrorsets
-          verbs:
-          - get
-          - list
-          - watch
-        - apiGroups:
           - ""
           resources:
           - configmaps
@@ -254,14 +244,6 @@ spec:
           - get
           - patch
           - update
-        - apiGroups:
-          - operator.openshift.io
-          resources:
-          - imagecontentsourcepolicies
-          verbs:
-          - get
-          - list
-          - watch
         - apiGroups:
           - rbac.authorization.k8s.io
           resources:
